@@ -11,13 +11,13 @@
         "php": "^7.1.3",
         "fideloper/proxy": "^4.0",
         "intervention/image": "2.4",
-        "laravel/framework": "^7.0",
+        "laravel/framework": "^8.0",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0"
+        "laravel/ui": "^3.0"
     },
     "require-dev": {
         "filp/whoops": "^2.0",
-        "nunomaduro/collision": "^4.1"
+        "nunomaduro/collision": "^5.0"
     },
     "config": {
         "optimize-autoloader": true,
@@ -31,12 +31,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
